+++ src/api/serverapi/serverapi.h
@@ -48,7 +48,8 @@ public:
     std::shared_ptr<WSNetCancelableCallback> confirmEmail(const std::string &authHash, WSNetRequestFinishedCallback callback) override;
     std::shared_ptr<WSNetCancelableCallback> signup(const std::string &username, const std::string &password, const std::string &referringUsername,
                                                     const std::string &email, const std::string &voucherCode, const std::string &secureToken, const std::string &captchaSolution,
-                                                    const std::vector<float> &captchaTrailX, const std::vector<float> &captchaTrailY, WSNetRequestFinishedCallback callback) override;
+                                                    const std::vector<float> &captchaTrailX, const std::vector<float> &captchaTrailY, const std::string &attestationToken,
+                                                    WSNetRequestFinishedCallback callback) override;
 
     std::shared_ptr<WSNetCancelableCallback> webSession(const std::string &authHash, WSNetRequestFinishedCallback callback) override;
     std::shared_ptr<WSNetCancelableCallback> checkUpdate(UpdateChannel updateChannel,
@@ -97,7 +98,8 @@ public:
                                                                        WSNetRequestFinishedCallback callback) override;
 
     std::shared_ptr<WSNetCancelableCallback> regToken(WSNetRequestFinishedCallback callback) override;
-    std::shared_ptr<WSNetCancelableCallback> signupUsingToken(const std::string &token, WSNetRequestFinishedCallback callback) override;
+    std::shared_ptr<WSNetCancelableCallback> signupUsingToken(const std::string &token, const std::string &attestationToken,
+                                                              WSNetRequestFinishedCallback callback) override;
     std::shared_ptr<WSNetCancelableCallback> claimAccount(const std::string &authHash, const std::string &username, const std::string &password,
                                                                   const std::string &email, const std::string &voucherCode, const std::string &claimAccount,
                                                                   WSNetRequestFinishedCallback callback) override;
@@ -113,7 +115,7 @@ public:
     std::shared_ptr<WSNetCancelableCallback> cancelAccount(const std::string &authHash, const std::string &password,
                                                                WSNetRequestFinishedCallback callback) override;
     std::shared_ptr<WSNetCancelableCallback> sso(const std::string &provider, const std::string &token,
-                                                           WSNetRequestFinishedCallback callback) override;
+                                                 const std::string &attestationToken, WSNetRequestFinishedCallback callback) override;
     std::shared_ptr<WSNetCancelableCallback> authTokenLogin(const std::string &username, bool useAsciiCaptcha, WSNetRequestFinishedCallback callback) override;
     std::shared_ptr<WSNetCancelableCallback> authTokenSignup(const std::string &username, bool useAsciiCaptcha, WSNetRequestFinishedCallback callback) override;
     std::shared_ptr<WSNetCancelableCallback> passwordRecovery(const std::string &email, WSNetRequestFinishedCallback callback) override;